add CTAS adapter unit tests
